Successfully implemented the GDT and the IDT. Changes to be committed:     modified:   src/arch/x86/boot.s 	modified:   src/gdt.c 	modified:   src/include/head.h 	modified:   src/kernel.c

+++ src/include/head.h
@@ -1,7 +1,7 @@
 #include <stdint.h>
 typedef unsigned char   uchar;
-typedef unsigned short  u16;
-typedef unsigned long   u32;
+typedef uint16_t        u16;
+typedef uint32_t        u32;
 typedef unsigned int    uint;
 typedef uint64_t        u64;
